v0.0.5 Release
Summary:
  Added pounds as weight unit
  Removed reauthentication auth log
  Added ingredient drag and drop reorder
  Added draggable icon
  Added inventory categories

+++ deploy/upgrade.sh
@@ -2,26 +2,31 @@ set -e
 sudo echo
 
 cd $(dirname $0)
-config_file="$(pwd)/config.env"
 
 # Git update
 cd ~/repos/Cardamom/
 git switch master; git fetch; git reset --hard origin/master
 
-source $config_file
+version=$(cat ~/repos/Cardamom/version.txt)
+source ~/repos/Cardamom/deploy/config.env
 echo "Upgrade Web: $upgrade_web"
 echo "Upgrade Database: $upgrade_database"
 echo "Upgrade API: $upgrade_api"
 
+read -p "Deploy version $version [y/N]: "
+if [[ !($REPLY =~ ^[Yy]$) ]]; then
+  echo "Aborting upgrade."
+  exit 1
+fi
+
 # Tag the version
-version=$(cat ~/repos/Cardamom/version.txt)
 git tag versions/$version -f
 git push --tags -f
 
 if [[ $upgrade_web == "yes" ]]; then
   # Upgrade web
   cd ~/repos/Cardamom/web
-  rm -r build/
+  rm -r build/ || echo "No build folder found"
   yarn build
   echo "Upgrading front end web application"
   sudo rm -rf /var/www/html/app.cardamom.cooking/*
@@ -33,7 +38,7 @@ fi
 # Upgrade API
 cd ~/repos/Cardamom/core
 
-if [[ $upgrade_db == "yes" ]]; then
+if [[ $upgrade_database == "yes" ]]; then
   # Database Migrate
   echo "Migrating database"
   go run bin/migration/migrate/migrate.go
